simple-web-service/requirements.txt
streamlit==1.35.0
faiss-cpu==1.8.0
transformers==4.41.2
sentence-transformers==3.0.0
torch==2.3.1
pymupdf==1.24.4
langchain==0.2.1
datasets==2.19.1  # 固定兼容版本
accelerate==0.30.1
modelscope==1.11.0